buttons: 32-button HID via MCP23017 I2C; fix -D tuning-flag forwarding
Buttons:
- HID descriptor button field 8 -> 32; wheel report `buttons` uint8 -> uint32.
- New buttons.c/.h: read momentary buttons from 1-2 MCP23017 I2C expanders
  (16/32 buttons on 2 wires). Internal pull-ups, ~12 ms debounce, scanned from
  the input-report path on core0. All I2C uses timeouts, so a missing chip
  degrades to "no buttons" instead of stalling the USB loop. Default disabled
  (BUTTON_CHIPS=0). Link hardware_i2c.

Build fix (latent, affected every knob):
- CMakeLists never forwarded `cmake -D<NAME>=<val>` to the compiler — it only
  set an unused CMake cache variable, so the C #ifndef default always won and
  MAX_NM / TORQUE_SIGN / PEDAL_COUNT / ... were NOT overridable as documented.
  Added a foreach block that forwards the defined knobs via
  target_compile_definitions. Verified -DBUTTON_CHIPS=2 -DPEDAL_COUNT=3 reach
  the compile commands and the enabled paths build clean.

Cleanup:
- Fix zero-length-array warning when PEDAL_COUNT==0 (size max(1,PEDAL_COUNT)).
- README: buttons section, tuning table rows, fix log 36-38.
- Refresh firmware/ffb_wheel.uf2 (RP2040, default config: buttons/pedals off).

Built clean with Pico SDK 2.1.1: default text 27640 / bss 7724;
with 2 chips + 3 pedals text 29860 / bss 7796.

+++ src/ffb.c
@@ -462,7 +462,7 @@ void ffb_calculate(const ffb_axis_metrics_t *m) {
 /* ============================================================ */
 
 uint16_t ffb_build_wheel_report(uint8_t *buffer, uint16_t reqlen,
-                                uint8_t buttons,
+                                uint32_t buttons,
                                 int16_t x, int16_t y, int16_t z,
                                 int16_t rx, int16_t ry, int16_t rz) {
     if (reqlen < sizeof(ffb_wheel_report_t)) return 0;
